Add bitsets

+++ skyspark/src/test/java/com/github/dkanellis/skyspark/api/test_utils/data_mocks/bitmap/FullBitmapStructureMock.java
@@ -10,7 +10,7 @@ public interface FullBitmapStructureMock {
 
     JavaRDD<Double> getDistinctValuesSorted();
 
-    JavaPairRDD<Long, Double> getValuesIndexed();
+    JavaPairRDD<Double, Long> getValuesIndexed();
 
     JavaRDD<BitSet> getValuesBitSets();
 }
